[kde-base/kde-l10n-*] Bumped to 4.4.3 and cleaned up old ebuilds.
Added ebuilds for kde l10n packages for KDE version 4.4.3, and
removed ebuilds for older versions of KDE where that version of KDE
was no longer available in portage.

+++ kde-base/do-kde-l10n.sh
@@ -8,13 +8,15 @@
 # EXISTING_VER:  An existing version to be used as a template
 # NEW_VER: The new version number to be generated.
 
-EXISTING_VER="4.4.1"
-NEW_VER="4.4.2"
+EXISTING_VER="4.4.2"
+NEW_VER="4.4.3"
 
 for X in `find -name kde-l10n-*${EXISTING_VER}*.ebuild`; do
         echo
         echo " ________ DOING "${X}" ________"
         echo
-        cp ${X} ${X/${EXISTING_VER}/${NEW_VER}}
+        if [ ! -e ${X/${EXISTING_VER}/${NEW_VER}} ]; then
+                cp ${X} ${X/${EXISTING_VER}/${NEW_VER}}
+        fi
         ebuild ${X} manifest
 done
